+++ cmd/goparamcount/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic
 Versioning].
 
+## [Unreleased]
+
+- _No changes yet_
+
 ## [1.0.0] - 2020-10-06
 
 - Add `-tests` flag.
@@ -36,7 +40,7 @@ Versioning].
 ## [0.1.0] - 2020-09-20
 
 - Analyse Go code to find function definitions with too many parameters.
-- Configure the number of parameters allowed (default is `2`).
+- Configure the number of parameters allowed ~~(default is `2`)~~.
 - Use the `-set_exit_status` flag to exit with a non-zero exit code if any
   function exceeds the limit.